+++ docs/CHANGELOG.md
@@ -10,6 +10,9 @@ Features:
 Bugfixes:
 - 
 
+Other:
+- Refactored Rare catches tracker to not execute any checks when setting is disabled / when being in a wrong world.
+
 ## v1.36.0
 
 Released: 2025-04-07